v0.9.7

+++ src/main/poke.c
@@ -1,5 +1,5 @@
 // 
-// Copyright (C) 2001 Graeme Walker <graeme_walker@users.sourceforge.net>
+// Copyright (C) 2001-2002 Graeme Walker <graeme_walker@users.sourceforge.net>
 // 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -82,7 +82,7 @@ int main( int argc , char * argv [] )
 
 	/* send the string */
 	rc = write( fd , buffer , strlen(buffer) ) ;
-	if( rc != strlen(buffer) ) 
+	if( rc != (int)strlen(buffer) ) 
 		return EXIT_FAILURE ;
 
 	/* read the reply */
